rules for self-references in structs
Make constructors do the renaming necessary to support self references.
There are probably some bugs around parameters, as we are not
necessarily tracking the renaming needed there. Is this a problem that
will occur in other places we do renames?
Should we be threading a substitution back that can be applied,
or perhaps threading *in* lists of parameters in the environment
that can be popped back off? Best to explore with some tests.