don't make is_ modify env
It's going to be useful when constructing
lien chains so that `our P` if `shared(P)`
can just be `P`.

This may cause challenges with inference later
but (a) we may never model inference and (b) even
if we do, we'd probably prefer to find a model
that didn't branch during lien chain
construction.