enumerate valid my-subtyping
Revisiting the subtyping rules, I am changing
the setup from having "my" be a subtype of
anything and then separately checking layout
compatibility to enumerating the cases
that my is a subtype of.

I'm not sure if this has unintended effects.
The previous structure I think was intended to
separate out "permission" subtyping from "layout
compatibility" and there is some appeal to that,
but overall this setup seems simpler.