introduce predicates, simplify and document rules
This is not obviously equivalent to the old rules.
I am not sure what I was trying to achieve
with sublienchain_exts. I need to better write up
the reasoning for why these rules are the way they
are. That said, all tests pass and it *seems*
correct to me.