WIP: rework red-perms
The old approach of breaking everything
down to individual chains was very flexible
but it was interacting poorly with inference.
It meant that the bounds on an inference
variable could were sets of sets of chains
and it was hard to figure out the permission
that they would result in.

This new approach aims to be simpler.
It also alters how things work a bit
to not be "recursive" in the types of borrowed
places (FIXME: I don't think we do this
everywhere we need to -- have to double check,
but for example I'm not convinced the handling
of dead places is correct).

Consider:

```
p: our Something
q: shared(p) String
```

in the new system, we don't rewrite the type of `q`
to be `our String`. In fact that will happen
at the point where `.share` occurs, since the
resulting type would have the surrounding
permission already applied. I can't come up with
an example right now that goes wrong,
but regardless it makes everything simpler.

We have intentionally simplified dead places;
a `shared(a, b, c)` prefix can only be removed
if all of `a`, `b`, and `c` are dead.
We might be able to do better -- and
even need to -- you can imagine
a case where you have `sh[a] String` on one
side and `sh[b] sh[a] String` or
some such thing, don't hae time to fully elaborate
the example, but basically on one side the
data comes directly from `a` and on the other
indirectly via `b`, and the question is, once
`b` is dead, can we do upcasting?